parsers and splitters moved according to the hierarchy
