+++ src/main/kotlin/astminer/examples/AllJavaMethods.kt
@@ -2,9 +2,9 @@ package astminer.examples
 
 import astminer.common.model.LabeledPathContexts
 import astminer.common.model.MethodInfo
-import astminer.parse.java.GumTreeJavaNode
-import astminer.parse.java.GumTreeJavaParser
-import astminer.parse.java.GumTreeJavaMethodSplitter
+import astminer.parse.gumtree.java.GumTreeJavaNode
+import astminer.parse.gumtree.java.GumTreeJavaParser
+import astminer.parse.gumtree.java.GumTreeJavaMethodSplitter
 import astminer.paths.PathMiner
 import astminer.paths.PathRetrievalSettings
 import astminer.paths.CsvPathStorage
