+++ src/main/kotlin/astminer/parse/fuzzy/cpp/utils.kt
@@ -1,4 +1,4 @@
-package astminer.parse.cpp
+package astminer.parse.fuzzy.cpp
 
 import java.io.File
 import java.util.concurrent.TimeUnit
