+++ src/main/kotlin/astminer/parse/gumtree/java/GumTreeJavaNode.kt
@@ -1,10 +1,10 @@
-package astminer.parse.python
+package astminer.parse.gumtree.java
 
-import astminer.common.model.Node
 import com.github.gumtreediff.tree.ITree
 import com.github.gumtreediff.tree.TreeContext
+import astminer.common.model.Node
 
-class GumTreePythonNode(val wrappedNode: ITree, val context: TreeContext, val parent: GumTreePythonNode?) : Node {
+class GumTreeJavaNode(val wrappedNode: ITree, val context: TreeContext, val parent: GumTreeJavaNode?) : Node {
     private val metadata: MutableMap<String, Any> = HashMap()
 
     override fun getMetadata(key: String): Any? {
@@ -19,8 +19,8 @@ class GumTreePythonNode(val wrappedNode: ITree, val context: TreeContext, val pa
         return childrenList.isEmpty()
     }
 
-    private val childrenList: MutableList<GumTreePythonNode> by lazy {
-        wrappedNode.children.map { GumTreePythonNode(it, context, this) }.toMutableList()
+    private val childrenList: MutableList<GumTreeJavaNode> by lazy {
+        wrappedNode.children.map { GumTreeJavaNode(it, context, this) }.toMutableList()
     }
 
     override fun getTypeLabel(): String {
@@ -42,4 +42,5 @@ class GumTreePythonNode(val wrappedNode: ITree, val context: TreeContext, val pa
     override fun removeChildrenOfType(typeLabel: String) {
         childrenList.removeIf { it.getTypeLabel() == typeLabel }
     }
+
 }
