+++ src/test/kotlin/astminer/parse/cpp/FuzzyCppParserTest.kt
@@ -2,6 +2,8 @@ package astminer.parse.cpp
 
 import astminer.common.getProjectFilesWithExtension
 import astminer.examples.forFilesWithSuffix
+import astminer.parse.fuzzy.cpp.FuzzyCppParser
+import astminer.parse.fuzzy.cpp.FuzzyNode
 import org.junit.Assert
 import org.junit.Test
 import java.io.File
